@@ -20,20 +20,19 @@ classifiers = [
   "Operating System :: OS Independent",
   "Programming Language :: Python",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
   "Programming Language :: Python :: 3.13",
+  "Programming Language :: Python :: 3.14",
   "Programming Language :: R",
   "Topic :: Scientific/Engineering",
   "Topic :: Scientific/Engineering :: Information Analysis",
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
   "click",
   "genno >= 1.20",
-  "importlib_metadata",
   "JPype1 >= 1.2.1",
   "JPype1 <= 1.4.0; python_version < '3.11'",
   "openpyxl",
@@ -49,7 +48,6 @@ Documentation = "https://docs.messageix.org/ixmp"
 Repository = "https://github.com/iiasa/ixmp"
 "Issue tracker" = "https://github.com/iiasa/ixmp/issues"
 
-
 [project.optional-dependencies]
 docs = [
   "ixmp[tests]",
@@ -61,7 +59,7 @@ docs = [
   "sphinxcontrib-bibtex",
 ]
 ixmp4 = [
-  "ixmp4 >= 0.13, < 0.14; python_version > '3.9'",
+  "ixmp4 >= 0.13, < 0.14",
   "gamsapi[core,transfer] >= 45.7.0",
 ]
 report = ["genno[compat,graphviz]"]
@@ -154,6 +152,7 @@ markers = [
   "performance: ixmp performance test.",
   "jdbc: tests exclusive to JDBCBackend.",
   "ixmp4: tests exclusive to IXMP4Backend.",
+  "ixmp4_209: https://github.com/iiasa/ixmp4/pull/209",
 ]
 tmp_path_retention_policy = "none"
 usefixtures = "tmp_env"
