@@ -5,7 +5,6 @@
 
 import importlib.metadata
 from pathlib import Path
-from typing import Optional
 
 # -- Project information ---------------------------------------------------------------
 
@@ -101,7 +100,7 @@ extlinks = {
 # -- Options for sphinx.ext.intersphinx ------------------------------------------------
 
 
-def local_inv(name: str, *parts: str) -> Optional[str]:
+def local_inv(name: str, *parts: str) -> str | None:
     """Construct the path to a local intersphinx inventory."""
 
     from importlib.util import find_spec
