@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Any, Optional
+from typing import TYPE_CHECKING, Any
 
 from .dantzig import DantzigModel
 from .gams import GAMSModel
@@ -15,7 +15,7 @@ MODELS: dict[str, type["Model"]] = {
 }
 
 
-def get_model(name: Optional[str], **model_options: Any) -> "Model":
+def get_model(name: str | None, **model_options: Any) -> "Model":
     """Return a model instance for `name` (or the default) with `model_options`.
 
     Note that unlike :func:`.backend.get_class`, this function creates a new instance.
