@@ -1,14 +1,14 @@
 from pathlib import Path
-from typing import Any, Literal, Union, cast
+from typing import TYPE_CHECKING, Any, Literal, cast
 
 import pandas as pd
 import pytest
 
 from ixmp import Scenario
 from ixmp.backend.common import ItemType
-from ixmp.testing import min_ixmp4_version
 
-pytestmark = min_ixmp4_version
+if TYPE_CHECKING:
+    from ixmp.backend.ixmp4 import IXMP4Backend
 
 
 def test__ensure_filters_values_are_lists() -> None:
@@ -49,11 +49,12 @@ class TestIxmp4Functions:
     # that's also used by JDBC, and is always called for Backend.init() and
     # Backend.get(), so is tested already.
 
-    def test__ni(self, ixmp4_backend: Any) -> Any:
+    def test__ni(self, ixmp4_backend: "IXMP4Backend") -> Any:
         with pytest.raises(NotImplementedError):
             ixmp4_backend._ni()
 
-    def test__get_repo(self, ixmp4_backend: Any, scenario: Scenario) -> None:
+    @pytest.mark.ixmp4_209
+    def test__get_repo(self, ixmp4_backend: "IXMP4Backend", scenario: Scenario) -> None:
         from ixmp4.core.optimization.equation import Equation, EquationRepository
         from ixmp4.core.optimization.indexset import IndexSet, IndexSetRepository
         from ixmp4.core.optimization.parameter import Parameter, ParameterRepository
@@ -72,10 +73,13 @@ class TestIxmp4Functions:
 
         # Test correct kind of instance is returned
         for type, expected_repo in repos.items():
-            repo = ixmp4_backend._get_repo(s=scenario, type=type)
+            repo = ixmp4_backend._get_repo(s=scenario, type=type)  # type: ignore [arg-type]
             assert isinstance(repo, expected_repo)
 
-    def test__find_item(self, ixmp4_backend: Any, scenario: Scenario) -> None:
+    @pytest.mark.ixmp4_209
+    def test__find_item(
+        self, ixmp4_backend: "IXMP4Backend", scenario: Scenario
+    ) -> None:
         from ixmp4.core.optimization.indexset import IndexSet
 
         # Test unknown item name raises
@@ -91,8 +95,9 @@ class TestIxmp4Functions:
         )
         assert (_type, name) == (type(return_item), return_item.name)
 
+    @pytest.mark.ixmp4_209
     def test__get_indexset_or_table(
-        self, ixmp4_backend: Any, scenario: Scenario
+        self, ixmp4_backend: "IXMP4Backend", scenario: Scenario
     ) -> None:
         # Test getting an item of "unknown" type (either 'indexset' or 'table');
         # getting a Table first tests getting an IndexSet, too
@@ -103,7 +108,10 @@ class TestIxmp4Functions:
         table = ixmp4_backend._get_indexset_or_table(s=scenario, name=table_name)
         assert table.name == table_name
 
-    def test__add_data_to_set(self, ixmp4_backend: Any, scenario: Scenario) -> None:
+    @pytest.mark.ixmp4_209
+    def test__add_data_to_set(
+        self, ixmp4_backend: "IXMP4Backend", scenario: Scenario
+    ) -> None:
         # Test adding to an Indexset and warning about `comment`
         indexset_name = "Indexset"
         scenario.init_set(name=indexset_name)
@@ -125,7 +133,10 @@ class TestIxmp4Functions:
             pd.DataFrame({indexset_name: [key]}),
         )
 
-    def test__create_scalar(self, ixmp4_backend: Any, scenario: Scenario) -> None:
+    @pytest.mark.ixmp4_209
+    def test__create_scalar(
+        self, ixmp4_backend: "IXMP4Backend", scenario: Scenario
+    ) -> None:
         # Test creating a scalar with a comment
         name = "Scalar"
         value = 3.141592653589793238462643383279502884197169399375105820
@@ -138,8 +149,9 @@ class TestIxmp4Functions:
         assert scalar.value == value
         assert scalar.docs == comment
 
+    @pytest.mark.ixmp4_209
     def test__add_data_to_parameter(
-        self, ixmp4_backend: Any, scenario: Scenario
+        self, ixmp4_backend: "IXMP4Backend", scenario: Scenario
     ) -> None:
         # Set up auxiliary items
         unit_name = "Unit"
@@ -164,7 +176,10 @@ class TestIxmp4Functions:
             "units": [unit_name],
         }
 
-    def test__get_set_data(self, ixmp4_backend: Any, scenario: Scenario) -> None:
+    @pytest.mark.ixmp4_209
+    def test__get_set_data(
+        self, ixmp4_backend: "IXMP4Backend", scenario: Scenario
+    ) -> None:
         indexset_name = "Indexset"
         scenario.init_set(name=indexset_name)
         table_name = "Table"
@@ -176,7 +191,7 @@ class TestIxmp4Functions:
         # TODO But we could first figure out if we're dealing with indexset or table and
         # overload _get_set_data to eliminate the need for cast()
         pd.testing.assert_series_equal(
-            cast("pd.Series[Union[float, int, str]]", indexset_data), pd.Series([])
+            cast("pd.Series[float | int | str]", indexset_data), pd.Series([])
         )
         table_data = ixmp4_backend._get_set_data(s=scenario, name=table_name)
         # We can assume this return type for Tables
@@ -185,7 +200,7 @@ class TestIxmp4Functions:
         )
 
     # Test some edge cases for standard functions
-    def test_handle_config(self, ixmp4_backend: Any) -> None:
+    def test_handle_config(self, ixmp4_backend: "IXMP4Backend") -> None:
         # Test raising for unhandled positional args
         with pytest.raises(ValueError, match="Unhandled positional args"):
             ixmp4_backend.handle_config(["test arg"], {"foo": "bar"})
@@ -195,7 +210,7 @@ class TestIxmp4Functions:
             ixmp4_backend.handle_config([], {"foo": "bar"})
 
     def test_set_node(
-        self, ixmp4_backend: Any, caplog: pytest.LogCaptureFixture
+        self, ixmp4_backend: "IXMP4Backend", caplog: pytest.LogCaptureFixture
     ) -> None:
         from ixmp.backend.ixmp4 import log
 
@@ -215,8 +230,12 @@ class TestIxmp4Functions:
         ]
         assert caplog.messages == expected
 
+    @pytest.mark.ixmp4_209
     def test_clone(
-        self, ixmp4_backend: Any, caplog: pytest.LogCaptureFixture, scenario: Scenario
+        self,
+        ixmp4_backend: "IXMP4Backend",
+        caplog: pytest.LogCaptureFixture,
+        scenario: Scenario,
     ) -> None:
         from ixmp.backend.ixmp4 import log
 
@@ -237,8 +256,12 @@ class TestIxmp4Functions:
         )
         assert expected in caplog.messages
 
+    @pytest.mark.ixmp4_209
     def test_clear_solution(
-        self, ixmp4_backend: Any, caplog: pytest.LogCaptureFixture, scenario: Scenario
+        self,
+        ixmp4_backend: "IXMP4Backend",
+        caplog: pytest.LogCaptureFixture,
+        scenario: Scenario,
     ) -> None:
         from ixmp.backend.ixmp4 import log
 
@@ -251,12 +274,16 @@ class TestIxmp4Functions:
         )
         assert expected in caplog.messages
 
-    def test_run_id(self, ixmp4_backend: Any, scenario: Scenario) -> None:
+    @pytest.mark.ixmp4_209
+    def test_run_id(self, ixmp4_backend: "IXMP4Backend", scenario: Scenario) -> None:
         # NOTE Depending on what run_id() should actually fetch, this needs adapting
         # scenario sets up a new Run, which has version 1
         assert ixmp4_backend.run_id(ts=scenario) == 1
 
-    def test_item_delete_elements(self, ixmp4_backend: Any, scenario: Scenario) -> None:
+    @pytest.mark.ixmp4_209
+    def test_item_delete_elements(
+        self, ixmp4_backend: "IXMP4Backend", scenario: Scenario
+    ) -> None:
         # Prepare some data
         run = ixmp4_backend.index[scenario]
         indexset_data = "foo"
@@ -282,7 +309,10 @@ class TestIxmp4Functions:
         assert isinstance(new_data, pd.DataFrame)
         assert new_data.empty
 
-    def test_delete_item(self, ixmp4_backend: Any, scenario: Scenario) -> None:
+    @pytest.mark.ixmp4_209
+    def test_delete_item(
+        self, ixmp4_backend: "IXMP4Backend", scenario: Scenario
+    ) -> None:
         # Create a 'set' to delete
         run = ixmp4_backend.index[scenario]
         with run.transact("Test delete_item"):
@@ -292,7 +322,7 @@ class TestIxmp4Functions:
         # Test there are no 'sets' on scenario anymore
         assert ixmp4_backend.list_items(s=scenario, type="set") == []
 
-    def test_write_file(self, ixmp4_backend: Any) -> None:
+    def test_write_file(self, ixmp4_backend: "IXMP4Backend") -> None:
         # Test raising an error for unknown file extension
         with pytest.raises(NotImplementedError):
             ixmp4_backend.write_file(
@@ -305,7 +335,8 @@ class TestIxmp4Functions:
                 path=Path("none.gdx"), item_type=ItemType.EQU, filters={}
             )
 
-    def test_read_file(self, ixmp4_backend: Any, scenario: Scenario) -> None:
+    @pytest.mark.ixmp4_209
+    def test_read_file(self, ixmp4_backend: "IXMP4Backend", scenario: Scenario) -> None:
         # Test raising an error for unknown file extension
         with pytest.raises(NotImplementedError):
             ixmp4_backend.read_file(
