@@ -1,21 +1,19 @@
 from pathlib import Path
-from typing import Any, Literal, Union
+from typing import TYPE_CHECKING, Any, Literal
 
 import pandas as pd
 import pytest
+from gams.transfer import Container
 
 from ixmp.model.gams import gams_info
-from ixmp.testing import min_ixmp4_version
 from ixmp.util.ixmp4 import ContainerData
 
-pytestmark = min_ixmp4_version
+if TYPE_CHECKING:
+    from ixmp4.core import Run
 
 
-# NOTE Generic return type for Python 3.9 compliance
 @pytest.fixture
-def container() -> Any:
-    from gams.transfer import Container
-
+def container() -> "Container":  # type: ignore [no-any-unimported]
     return Container(system_directory=str(gams_info().system_dir))
 
 
@@ -33,7 +31,7 @@ def test__record_versions(container: Any) -> None:
     # Test recording the package version of something that's never present
     _record_versions(container=container, packages=["not installed"])
     indexset = container.getSets()[0]
-    print(indexset.records)
+    # print(indexset.records)  # DEBUG
     pd.testing.assert_frame_equal(
         indexset.records[["uni_0", "uni_1"]],
         pd.DataFrame(
@@ -100,10 +98,11 @@ def backend() -> Literal["ixmp4"]:
 # NOTE This marker usually parametrizes test creation, but with overriding `backend`,
 # this might not be necessary. Keeping it for clarity.
 @pytest.mark.ixmp4
+@pytest.mark.ixmp4_209
 class TestIxmp4IOFunctions:
     """Test group for all IO functions touching ixmp4 directly."""
 
-    def test__domain(self, run: Any) -> None:
+    def test__domain(self, run: "Run") -> None:
         from ixmp.backend.ixmp4_io import _domain
 
         with run.transact("Test _domain"):
@@ -115,7 +114,7 @@ class TestIxmp4IOFunctions:
         assert _domain(indexset) is None
         assert _domain(table) == [indexset.name]
 
-    def test__records(self, run: Any) -> None:
+    def test__records(self, run: "Run") -> None:
         from ixmp.backend.ixmp4_io import _records
 
         run.backend.units.create("unit")
@@ -150,7 +149,7 @@ class TestIxmp4IOFunctions:
         expected.pop("units")
         assert _records(parameter) == expected
 
-    def test__ensure_correct_item_order(self, run: Any) -> None:
+    def test__ensure_correct_item_order(self, run: "Run") -> None:
         from ixmp.backend.ixmp4_io import _ensure_correct_item_order
 
         with run.transact("Test _ensure_correct_item_order"):
@@ -164,15 +163,13 @@ class TestIxmp4IOFunctions:
             repo=run.optimization.indexsets,
         ) == [years, nodes, type_years, type_nodes]
 
-    def test__align_records_and_domain(self, run: Any) -> None:
+    def test__align_records_and_domain(self, run: "Run") -> None:
         from ixmp.backend.ixmp4_io import _align_records_and_domain
 
         # Test item without domain order
         with run.transact("Test _align_records_and_domain without order"):
             equation = run.optimization.equations.create("Equation")
-        expected: dict[str, Union[list[float], list[int], list[str]]] = {
-            "foo": [1, 2, 3]
-        }
+        expected: dict[str, list[float] | list[int] | list[str]] = {"foo": [1, 2, 3]}
         assert _align_records_and_domain(item=equation, records=expected) == expected
 
         # Test sorting of records
@@ -182,7 +179,7 @@ class TestIxmp4IOFunctions:
             parameter = run.optimization.parameters.create(
                 "Parameter", constrained_to_indexsets=[indexset_1.name, indexset_2.name]
             )
-        records: dict[str, Union[list[float], list[int], list[str]]] = {
+        records: dict[str, list[float] | list[int] | list[str]] = {
             indexset_2.name: [1, 2, 3],
             "values": [1.0, 2.0, 3.0],
             indexset_1.name: ["1", "2", "3"],
@@ -195,7 +192,7 @@ class TestIxmp4IOFunctions:
 
         assert _align_records_and_domain(item=parameter, records=records) == expected
 
-    def test__convert_ixmp4_items_to_containerdata(self, run: Any) -> None:
+    def test__convert_ixmp4_items_to_containerdata(self, run: "Run") -> None:
         from ixmp.backend.ixmp4_io import _convert_ixmp4_items_to_containerdata
 
         # Test handling of empty list
@@ -223,7 +220,7 @@ class TestIxmp4IOFunctions:
         assert table.name == table_data.name
         assert table.data == table_data.records
 
-    def test_write_run_to_gdx(self, run: Any, tmp_path: Path) -> None:
+    def test_write_run_to_gdx(self, run: "Run", tmp_path: Path) -> None:
         from gams.transfer import Container
 
         from ixmp.backend.ixmp4_io import write_run_to_gdx
@@ -246,7 +243,7 @@ class TestIxmp4IOFunctions:
         assert len(container.data) == 1
         assert "ixmp_version" in container.data.keys()
 
-    def test__set_columns_to_read_from_records(self, run: Any) -> None:
+    def test__set_columns_to_read_from_records(self, run: "Run") -> None:
         from ixmp.backend.ixmp4_io import _set_columns_to_read_from_records
 
         default_columns = ["levels", "marginals"]
@@ -282,7 +279,7 @@ class TestIxmp4IOFunctions:
     # NOTE Keeping the read_equ()/read_var() tests separate because the read_*()
     # functions are separate
 
-    def test__read_variables_to_run(self, run: Any, container: Any) -> None:
+    def test__read_variables_to_run(self, run: "Run", container: Any) -> None:
         from ixmp.backend.ixmp4_io import _read_variables_to_run
 
         with run.transact("Test _read_variables_to_run"):
@@ -315,7 +312,7 @@ class TestIxmp4IOFunctions:
         )
         assert variable.data == expected
 
-    def test__read_equations_to_run(self, run: Any, container: Any) -> None:
+    def test__read_equations_to_run(self, run: "Run", container: Any) -> None:
         from ixmp.backend.ixmp4_io import _read_equations_to_run
 
         with run.transact("Test _read_equations_to_run"):
@@ -350,7 +347,7 @@ class TestIxmp4IOFunctions:
         )
         assert equation.data == expected
 
-    def test_read_gdx_to_run(self, run: Any, tmp_path: Path) -> None:
+    def test_read_gdx_to_run(self, run: "Run", tmp_path: Path) -> None:
         from ixmp.backend.ixmp4_io import read_gdx_to_run, write_run_to_gdx
 
         # NOTE Names without space to produce "valid GAMS names"
@@ -363,7 +360,7 @@ class TestIxmp4IOFunctions:
         variable_3 = run.backend.optimization.variables.create(
             run_id=run.id, name="Variable3"
         )
-        records: dict[str, Union[list[float], list[int], list[str]]] = {
+        records: dict[str, list[float] | list[int] | list[str]] = {
             "level": [1.0],
             "marginal": [0],
             "lower": [0],
