@@ -3,7 +3,7 @@
 #      behaviour they actually test.
 
 import copy
-from typing import Any, Generator, Union
+from typing import Any, Generator
 
 import pytest
 
@@ -49,7 +49,7 @@ class TestMeta:
     def test_set_meta_missing_argument(
         self,
         mp: ixmp.Platform,
-        meta: dict[str, Union[bool, int, str, list[Union[bool, float, int, str]]]],
+        meta: dict[str, bool | int | str | list[bool | float | int | str]],
     ) -> None:
         with pytest.raises(ValueError):
             mp.set_meta(meta)
@@ -62,7 +62,7 @@ class TestMeta:
     def test_set_get_meta(
         self,
         mp: ixmp.Platform,
-        meta: dict[str, Union[bool, int, str, list[Union[bool, float, int, str]]]],
+        meta: dict[str, bool | int | str | list[bool | float | int | str]],
     ) -> None:
         """Assert that storing+retrieving meta yields expected values."""
         mp.set_meta(meta, model=DANTZIG["model"])
@@ -73,7 +73,7 @@ class TestMeta:
     def test_unique_meta(
         self,
         mp: ixmp.Platform,
-        meta: dict[str, Union[bool, int, str, list[Union[bool, float, int, str]]]],
+        meta: dict[str, bool | int | str | list[bool | float | int | str]],
     ) -> None:
         """
         When setting a meta category on two distinct levels, a uniqueness error is
@@ -106,7 +106,7 @@ class TestMeta:
     def test_set_get_meta_equals(
         self,
         mp: ixmp.Platform,
-        meta: dict[str, Union[bool, int, str, list[Union[bool, float, int, str]]]],
+        meta: dict[str, bool | int | str | list[bool | float | int | str]],
     ) -> None:
         initial_meta = mp.get_meta(scenario=DANTZIG["scenario"])
         mp.set_meta(meta, model=DANTZIG["model"])
@@ -117,7 +117,7 @@ class TestMeta:
     def test_unique_meta_model_scenario(
         self,
         mp: ixmp.Platform,
-        meta: dict[str, Union[bool, int, str, list[Union[bool, float, int, str]]]],
+        meta: dict[str, bool | int | str | list[bool | float | int | str]],
     ) -> None:
         """
         When setting a meta key for a Model, it shouldn't be possible to set it
@@ -142,19 +142,19 @@ class TestMeta:
     def test_get_meta_strict(
         self,
         mp: ixmp.Platform,
-        meta: dict[str, Union[bool, int, str, list[Union[bool, float, int, str]]]],
+        meta: dict[str, bool | int | str | list[bool | float | int | str]],
     ) -> None:
         """
         Set meta indicators on several model/scenario/version levels and test
         the 'strict' parameter of get_meta().
         """
         # set meta on various levels
-        model_meta: dict[str, Union[bool, int, str]] = {
+        model_meta: dict[str, bool | int | str] = {
             "model_int": 3,
             "model_string": "string_value",
             "model_bool": False,
         }
-        scenario_meta: dict[str, Union[bool, int, str]] = {
+        scenario_meta: dict[str, bool | int | str] = {
             "scenario_int": 3,
             "scenario_string": "string_value",
             "scenario_bool": False,
@@ -246,7 +246,7 @@ class TestMeta:
     def test_unique_meta_scenario(
         self,
         mp: ixmp.Platform,
-        meta: dict[str, Union[bool, int, str, list[Union[bool, float, int, str]]]],
+        meta: dict[str, bool | int | str | list[bool | float | int | str]],
     ) -> None:
         """
         When setting a meta key on a specific Scenario run, setting the same key
