@@ -2,7 +2,7 @@ import logging
 import re
 from collections.abc import Generator
 from datetime import datetime, timedelta
-from typing import TYPE_CHECKING, Any, Optional, Union
+from typing import TYPE_CHECKING, Any
 
 import pandas as pd
 import pytest
@@ -54,8 +54,8 @@ def assert_geodata(obs: pd.DataFrame, exp: pd.DataFrame) -> None:
 def assert_timeseries(
     scen: TimeSeries,
     exp: pd.DataFrame = DATA["timeseries"],
-    cols: Optional[Union[str, list[str]]] = None,
-    subannual: Optional[Union[bool, str]] = None,
+    cols: str | list[str] | None = None,
+    subannual: bool | str | None = None,
 ) -> None:
     """Asserts scenario timeseries are similar to expected
 
@@ -84,6 +84,7 @@ def assert_timeseries(
 # Fixtures
 
 
+@pytest.mark.ixmp4_209
 class TestTimeSeries:
     """Tests of :class:`.TimeSeries`.
 
@@ -96,7 +97,7 @@ class TestTimeSeries:
     @pytest.fixture(scope="function", params=[TimeSeries, Scenario])
     def ts(
         self, request: pytest.FixtureRequest, mp: "Platform"
-    ) -> Generator[Union[TimeSeries, Scenario], Any, None]:
+    ) -> Generator[TimeSeries | Scenario, Any, None]:
         """An empty TimeSeries with a temporary name on the :func:`mp`."""
         # Use a hash of the pytest node ID to avoid exceeding the maximum
         # length for a scenario name
@@ -108,7 +109,7 @@ class TestTimeSeries:
     # Initialize TimeSeries
     @pytest.mark.parametrize("cls", [TimeSeries, Scenario])
     def test_init(
-        self, test_mp: "Platform", cls: Union[type[TimeSeries], type[Scenario]]
+        self, test_mp: "Platform", cls: type[TimeSeries] | type[Scenario]
     ) -> None:
         # Something other than a Platform as mp argument
         with pytest.raises(TypeError):
@@ -202,9 +203,6 @@ class TestTimeSeries:
 
         assert 0 == len(ts.timeseries())
 
-    # FIXME IXMP4Backend seems to handle timeseries() incorrectly; the last call returns
-    # 0 rows
-    @pytest.mark.jdbc
     @pytest.mark.parametrize("format", ["long", "wide"])
     def test_get(self, ts: TimeSeries, format: str) -> None:
         data = DATA[0] if format == "long" else wide(DATA[0])
@@ -236,7 +234,7 @@ class TestTimeSeries:
         ],
     )
     def test_get_year(
-        self, ts: TimeSeries, year_arg: Union[int, list[int], list[str]]
+        self, ts: TimeSeries, year_arg: int | list[int] | list[str]
     ) -> None:
         """`year` arg to :meth:`.TimeSeries.timeseries` accepts only :class:`int`."""
         ts.add_timeseries(DATA[0])
@@ -246,9 +244,6 @@ class TestTimeSeries:
         # NOTE Triggering this error on purpose
         ts.timeseries(year=year_arg)  # type: ignore[arg-type]
 
-    # FIXME IXMP4Backend seems to handle timeseries() incorrectly; the last call returns
-    # 0 rows
-    @pytest.mark.jdbc
     @pytest.mark.parametrize("format", ["long", "wide"])
     def test_edit(self, mp: "Platform", ts: TimeSeries, format: str) -> None:
         """Tests that data can be overwritten."""
@@ -293,7 +288,7 @@ class TestTimeSeries:
         self,
         mp: "Platform",
         ts: TimeSeries,
-        cls: Union[type[TimeSeries], type[Scenario]],
+        cls: type[TimeSeries] | type[Scenario],
     ) -> None:
         info: "TimeSeriesIdentifiers" = dict(model=ts.model, scenario=ts.scenario)
 
@@ -483,8 +478,6 @@ class TestTimeSeries:
         ts.commit("")
         assert "Dropped extra column(s) ['climate_model'] from data" in caplog.messages
 
-    # NOTE Not yet implemented on IXMP4Backend
-    @pytest.mark.jdbc
     def test_new_timeseries_as_iamc(self, test_mp: "Platform") -> None:
         # TODO rescue use of subannual= here
 
