@@ -6,13 +6,39 @@ Imports from this module **should** only occur within a :py:`if TYPE_CHECKING:` 
 import os
 from collections.abc import Mapping, Sequence
 from pathlib import Path
-from typing import Any, Literal, Optional, TypedDict, Union
+from typing import Any, Literal, TypedDict
 
 import pandas
+from ixmp4.core.optimization.equation import Equation, EquationRepository
+from ixmp4.core.optimization.indexset import IndexSet, IndexSetRepository
+from ixmp4.core.optimization.parameter import Parameter, ParameterRepository
+from ixmp4.core.optimization.scalar import Scalar, ScalarRepository
+from ixmp4.core.optimization.table import Table, TableRepository
+from ixmp4.core.optimization.variable import Variable, VariableRepository
+from ixmp4.data.abstract.optimization.equation import (
+    EquationRepository as BEEquationRepository,
+)
+from ixmp4.data.abstract.optimization.indexset import (
+    IndexSetRepository as BEIndexSetRepository,
+)
+from ixmp4.data.abstract.optimization.parameter import (
+    ParameterRepository as BEParameterRepository,
+)
+from ixmp4.data.abstract.optimization.scalar import (
+    ScalarRepository as BEScalarRepository,
+)
+from ixmp4.data.abstract.optimization.table import (
+    TableRepository as BETableRepository,
+)
+from ixmp4.data.abstract.optimization.variable import (
+    VariableRepository as BEVariableRepository,
+)
 
 # Compatibility with Python 3.11 and earlier
 # TODO Use "from typing import NotRequired" when dropping support for Python 3.10
 # TODO Use "type x = ..." instead of TypeAlias when dropping support for Python 3.11
+# TODO Import this from typing when dropping Python 3.11
+# TODO Use type x = ... instead of TypeAlias when dropping support for Python 3.11
 from typing_extensions import NotRequired, TypeAlias
 
 from ixmp.backend.common import ItemType
@@ -20,14 +46,40 @@ from ixmp.core.scenario import Scenario
 from ixmp.core.timeseries import TimeSeries
 from ixmp.util.ixmp4 import ContainerData
 
+IXMP4BackendRepository = (
+    BEIndexSetRepository
+    | BEEquationRepository
+    | BEParameterRepository
+    | BEScalarRepository
+    | BETableRepository
+    | BEVariableRepository
+)
+
+IXMP4Repository = (
+    IndexSetRepository
+    | EquationRepository
+    | ParameterRepository
+    | ScalarRepository
+    | TableRepository
+    | VariableRepository
+)
+
+#: Instances of IXMP4 types that contain model/scenario data.
+IXMP4ModelData = Equation | IndexSet | Parameter | Scalar | Table | Variable
+
+#: Classes of `IXMP4ModelData`, i.e. references to the classes themselves and not
+#: particular instances.
+IXMP4ModelDataType = (
+    type[Equation]
+    | type[IndexSet]
+    | type[Parameter]
+    | type[Scalar]
+    | type[Table]
+    | type[Variable]
+)
 #: Filters arguments to many functions. Generally non-str elements are converted to
 #: str(). Since object.__str__() exists, any Python class has a string representation.
-Filters: TypeAlias = Optional[
-    Union[
-        Mapping[str, Union[Any, Sequence[Any]]],
-        Mapping[str, Any],
-    ]
-]
+Filters: TypeAlias = Mapping[str, Any | Sequence[Any]] | Mapping[str, Any] | None
 
 #: Any of the members of :attr:`ItemType.MODEL`.
 ModelItemType: TypeAlias = Literal[
@@ -35,10 +87,10 @@ ModelItemType: TypeAlias = Literal[
 ]
 
 #: Return type of :meth:`.Scenario.set` for a simple/non-indexed set.
-SimpleSetData: TypeAlias = "pandas.Series[Union[float, int, str]]"
+SimpleSetData: TypeAlias = "pandas.Series[float | int | str]"
 
 #: Return type of :meth:`.Scenario.set`.
-SetData: TypeAlias = Union[SimpleSetData, pandas.DataFrame]
+SetData: TypeAlias = SimpleSetData | pandas.DataFrame
 
 
 class ScalarParData(TypedDict):
@@ -49,7 +101,7 @@ class ScalarParData(TypedDict):
 
 
 #: Return type of :meth:`.Scenario.par`.
-ParData: TypeAlias = Union[ScalarParData, pandas.DataFrame]
+ParData: TypeAlias = ScalarParData | pandas.DataFrame
 
 
 class ScalarSolutionData(TypedDict):
@@ -60,27 +112,27 @@ class ScalarSolutionData(TypedDict):
 
 
 #: Return type of :meth:`.Scenario.equ` and :meth:`.Scenario.var`.
-SolutionData: TypeAlias = Union[ScalarSolutionData, pandas.DataFrame]
+SolutionData: TypeAlias = ScalarSolutionData | pandas.DataFrame
 
 #: Valid values of the :py:`version=...` keyword argument to :class:`.TimeSeries` and
 #: :class:`.Scenario`.
-VersionType: TypeAlias = Optional[Union[int, Literal["new"]]]
+VersionType: TypeAlias = int | Literal["new"] | None
 
 
 #: Backend-related arguments to :class:`.Platform`.
 BackendInitKwargs = TypedDict(
-    "BackendInitKwargs", {"class": NotRequired[Union[Literal["jdbc", "ixmp4"], str]]}
+    "BackendInitKwargs", {"class": NotRequired[Literal["jdbc", "ixmp4"] | str]}
 )
 
 
 class JDBCBackendInitKwargs(TypedDict):
     """Keyword arguments to :class:`.JDBCBackend`."""
 
-    driver: NotRequired[Optional[str]]
-    path: NotRequired[Optional[Union[str, Path]]]
-    url: NotRequired[Optional[str]]
-    user: NotRequired[Optional[str]]
-    password: NotRequired[Optional[str]]
+    driver: NotRequired[str | None]
+    path: NotRequired[str | Path | None]
+    url: NotRequired[str | None]
+    user: NotRequired[str | None]
+    password: NotRequired[str | None]
 
 
 class IXMP4BackendInitKwargs(TypedDict):
@@ -88,7 +140,7 @@ class IXMP4BackendInitKwargs(TypedDict):
 
     ixmp4_name: str
     dsn: NotRequired[str]
-    jdbc_compat: NotRequired[Union[bool, str]]
+    jdbc_compat: NotRequired[bool | str]
 
 
 class PlatformInitKwargs(BackendInitKwargs, JDBCBackendInitKwargs):
@@ -98,21 +150,21 @@ class PlatformInitKwargs(BackendInitKwargs, JDBCBackendInitKwargs):
     :class:`IXMP4BackendInitKwargs`.
     """
 
-    jvmargs: NotRequired[Optional[Union[str, list[str]]]]
-    dbprops: NotRequired[Optional[os.PathLike[str]]]
+    jvmargs: NotRequired[str | list[str] | None]
+    dbprops: NotRequired[os.PathLike[str] | None]
     cache: NotRequired[bool]
-    log_level: NotRequired[Optional[Union[int, str]]]
+    log_level: NotRequired[int | str | None]
     ixmp4_name: NotRequired[str]
     dsn: NotRequired[str]
-    jdbc_compat: NotRequired[Union[bool, str]]
+    jdbc_compat: NotRequired[bool | str]
 
 
 class InitializeItemsKwargs(TypedDict):
     """Keyword arguments to :meth:`.base.Model.initialize_items`."""
 
     ix_type: str
-    idx_sets: NotRequired[Optional[Sequence[str]]]
-    idx_names: NotRequired[Optional[Sequence[str]]]
+    idx_sets: NotRequired[Sequence[str] | None]
+    idx_names: NotRequired[Sequence[str] | None]
 
 
 class GamsModelInitKwargs(TypedDict, total=False):
@@ -125,9 +177,9 @@ class GamsModelInitKwargs(TypedDict, total=False):
     solve_args: list[str]
     gams_args: list[str]
     check_solution: bool
-    comment: Optional[str]
-    equ_list: Optional[list[str]]
-    var_list: Optional[list[str]]
+    comment: str | None
+    equ_list: list[str] | None
+    var_list: list[str] | None
     quiet: bool
     use_temp_dir: bool
     record_version_packages: Sequence[str]
@@ -190,15 +242,15 @@ class WriteKwargs(TypedDict, total=False):
     filters: WriteFilters
     record_version_packages: Sequence[str]
     container_data: list["ContainerData"]
-    max_row: Optional[int]
+    max_row: int | None
 
 
 class ReadKwargs(TypedDict, total=False):
     """Keyword arguments to :meth:`.Backend.read_file`."""
 
     filters: WriteFilters
-    firstyear: Optional[int]
-    lastyear: Optional[int]
+    firstyear: int | None
+    lastyear: int | None
     add_units: bool
     init_items: bool
     commit_steps: bool
